refactor(fiat-onramp): rename form bloc events and use Decimal
- bloc events and states were not in line with bloc naming conventions
- double was still being used in places instead of decimal

both changes were made with new&old refactoring efforts in mind

+++ lib/bloc/fiat/fiat_onramp_form/fiat_form_state.dart
@@ -90,10 +90,10 @@ final class FiatFormState extends Equatable with FormzMixin {
       selectedPaymentMethod.transactionLimits.firstOrNull;
 
   /// The minimum fiat amount that is allowed for the selected payment method
-  double? get minFiatAmount => transactionLimit?.min;
+  Decimal? get minFiatAmount => transactionLimit?.min;
 
   /// The maximum fiat amount that is allowed for the selected payment method
-  double? get maxFiatAmount => transactionLimit?.max;
+  Decimal? get maxFiatAmount => transactionLimit?.max;
   bool get isLoadingCurrencies => fiatList.length < 2 || coinList.length < 2;
   bool get isLoading => isLoadingCurrencies || status == FiatFormStatus.loading;
   bool get canSubmit =>
